+++ lib/src/models/enums.dart
@@ -1,2 +1,0 @@
-enum ProductType { consumable, nonConsumable, subscription }
-enum PurchaseStatusU { pending, purchased, restored, canceled, expired, error }