+++ lib/src/models/models.dart
@@ -1,4 +1,3 @@
 export 'entitlement.dart';
-export 'enums.dart';
 export 'product.dart';
 export 'purchase.dart';